feat(infra): Adjusted GitHub workflows for new infrastructure project location
Remove unnecessary project configuration files from repo

Deleted several .idea configuration files and a few unnecessary project files as they were not needed in the repository. This clean-up helps in maintaining a cleaner codebase and a more efficient repository by reducing clutter and the size of the repo. These files are usually user-specific and can be locally generated and should not be part of the central project repository.